Rename platforms to platform/targets
